@@ -1,8 +1,8 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 24
-    buildToolsVersion "29.0.2"
+    compileSdkVersion 30
+    buildToolsVersion "30.0.3"
 
     defaultConfig {
         applicationId "com.tencent.ncnnyolox"
@@ -13,12 +13,11 @@ android {
 
     externalNativeBuild {
         cmake {
-            version "3.10.2"
             path file('src/main/jni/CMakeLists.txt')
         }
     }
 
     dependencies {
-        implementation 'com.android.support:support-v4:24.0.0'
+        // implementation 'com.android.support:support-v4:28.0.0'
     }
 }
