@@ -20,7 +20,6 @@ import android.content.pm.PackageManager;
 import android.graphics.PixelFormat;
 import android.os.Bundle;
 import android.util.Log;
-import android.view.Surface;
 import android.view.SurfaceHolder;
 import android.view.SurfaceView;
 import android.view.View;
@@ -29,22 +28,16 @@ import android.widget.AdapterView;
 import android.widget.Button;
 import android.widget.Spinner;
 
-import android.support.v4.app.ActivityCompat;
-import android.support.v4.content.ContextCompat;
 
 public class MainActivity extends Activity implements SurfaceHolder.Callback
 {
     public static final int REQUEST_CAMERA = 100;
 
-    private NcnnYolox ncnnyolox = new NcnnYolox();
+    private final NcnnYolox ncnnyolox = new NcnnYolox();
     private int facing = 0;
 
-    private Spinner spinnerModel;
-    private Spinner spinnerCPUGPU;
     private int current_model = 0;
     private int current_cpugpu = 0;
-
-    private SurfaceView cameraView;
 
     /** Called when the activity is first created. */
     @Override
@@ -55,27 +48,24 @@ public class MainActivity extends Activity implements SurfaceHolder.Callback
 
         getWindow().addFlags(WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON);
 
-        cameraView = (SurfaceView) findViewById(R.id.cameraview);
+        SurfaceView cameraView = (SurfaceView) findViewById(R.id.cameraview);
 
         cameraView.getHolder().setFormat(PixelFormat.RGBA_8888);
         cameraView.getHolder().addCallback(this);
 
         Button buttonSwitchCamera = (Button) findViewById(R.id.buttonSwitchCamera);
-        buttonSwitchCamera.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View arg0) {
+        buttonSwitchCamera.setOnClickListener(arg0 -> {
 
-                int new_facing = 1 - facing;
+            int new_facing = 1 - facing;
 
-                ncnnyolox.closeCamera();
+            ncnnyolox.closeCamera();
 
-                ncnnyolox.openCamera(new_facing);
+            ncnnyolox.openCamera(new_facing);
 
-                facing = new_facing;
-            }
+            facing = new_facing;
         });
 
-        spinnerModel = (Spinner) findViewById(R.id.spinnerModel);
+        Spinner spinnerModel = (Spinner) findViewById(R.id.spinnerModel);
         spinnerModel.setOnItemSelectedListener(new AdapterView.OnItemSelectedListener() {
             @Override
             public void onItemSelected(AdapterView<?> arg0, View arg1, int position, long id)
@@ -93,7 +83,7 @@ public class MainActivity extends Activity implements SurfaceHolder.Callback
             }
         });
 
-        spinnerCPUGPU = (Spinner) findViewById(R.id.spinnerCPUGPU);
+        Spinner spinnerCPUGPU = (Spinner) findViewById(R.id.spinnerCPUGPU);
         spinnerCPUGPU.setOnItemSelectedListener(new AdapterView.OnItemSelectedListener() {
             @Override
             public void onItemSelected(AdapterView<?> arg0, View arg1, int position, long id)
@@ -144,9 +134,9 @@ public class MainActivity extends Activity implements SurfaceHolder.Callback
     {
         super.onResume();
 
-        if (ContextCompat.checkSelfPermission(getApplicationContext(), Manifest.permission.CAMERA) == PackageManager.PERMISSION_DENIED)
+        if (checkSelfPermission(Manifest.permission.CAMERA) == PackageManager.PERMISSION_DENIED)
         {
-            ActivityCompat.requestPermissions(this, new String[] {Manifest.permission.CAMERA}, REQUEST_CAMERA);
+            requestPermissions(new String[] {Manifest.permission.CAMERA}, REQUEST_CAMERA);
         }
 
         ncnnyolox.openCamera(facing);
